% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degradeSITES.R
\name{degradeSITES}
\alias{degradeSITES}
\title{Gradually degrade sites}
\usage{
degradeSITES(network, sites, degrade = c("B", "STP", "NB"), mindeg = 1,
  maxdeg = 10, toplot = TRUE, pop = 1e+05, returnNET = TRUE)
}
\arguments{
\item{network}{adjacency network describing where birds are going}

\item{sites}{network sites including latitude and longitude, and population count at each site}

\item{degrade}{sites to degrade.Typically `c( "B", "STP", "NB")`. Can be B for breeding, STP for stopover sites, and/or NB for non-breeding residency}

\item{mindeg}{minimum percentage to degrade the site by (i.e. reduce carrying capacity by 1 percent)}

\item{maxdeg}{minimum percentage to degrade the site by (i.e. reduce carrying capacity by 10 percent)}

\item{toplot}{TRUE/FALSE to determine whether the output is plotted or not}

\item{pop}{number of individuals in the population}

\item{returnNET}{TRUE/FALSE whether or not the network is a return network or not}
}
\value{
a list containting the prioritisation (as a list of the sites in the order they are removed), the network which was randomly generated,
the tracks that were randomly generated, and the sites that were randomly generated for animals to use.
}
\description{
Degrate sites gradually
}
\examples{
# with a return network
net <- randomPARALLEL(nwintering=5,mean_dist=0, sd_dist=1500, nbreeding=5)
network <- shortenNET(net$network, from = "Ssupersink", to = "Nsupersink")

colnames(network)[1]<- rownames(network)[1] <- "Ssupersource"
colnames(network)[length(colnames(network))] <-  "Nsupersink"
rownames(network)[length(rownames(network))] <- "Nsupersink"
sites <- net$sites
degradeSITES(network,sites)

}
