% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migflow.R
\name{migflow}
\alias{migflow}
\title{migflow}
\description{
This package calculates the maximum flow through a network using integer linear programming.
}
\author{
Kiran Dhanjal-Adams \email{kiran.dhanjal.adams@gmail.com}
}
\keyword{"_PACKAGE"}
\keyword{internal}
