% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomCENTRAL.R
\name{randomCENTRAL}
\alias{randomCENTRAL}
\title{Generate a random network}
\usage{
randomCENTRAL(nbreeding = 10, nwintering = 10, nstop = 30,
  pop = 1e+05, toplot = TRUE)
}
\arguments{
\item{nbreeding}{Number of breeding sites}

\item{nwintering}{Number of wintering sites}

\item{nstop}{Number of stopover sites (shared during north and south migration)}

\item{pop}{population size flowing through network}

\item{toplot}{TRUE/FALSE to determine whether the output is plotted or not}
}
\value{
a list containting the network which was randomly generated,
the tracks that were randomly generated, and the sites that were randomly generated for animals to use.
}
\description{
This function generates a random migratory network
}
