% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomDIRECTED.R
\name{randomDIRECTED}
\alias{randomDIRECTED}
\title{Generate a random network}
\usage{
randomDIRECTED(nbreeding = 3, nwintering = 4, nstop = 10,
  pop = 1e+05, mean_dist = 1000, sd_dist = 200, toplot = TRUE)
}
\arguments{
\item{nbreeding}{Number of breeding sites}

\item{nwintering}{Number of wintering sites}

\item{nstop}{Number of stopover sites (shared during north and south migration)}

\item{pop}{population size flowing through network}

\item{mean_dist}{average distance the species can travel (this can be estimated from real data)}

\item{sd_dist}{standard deviation of the distances a species can travel (this can be estimated from real data)}

\item{toplot}{TRUE/FALSE to determine whether the output is plotted or not}
}
\value{
a list containting the network which was randomly generated,
the tracks that were randomly generated, and the sites that were randomly generated for animals to use.
}
\description{
This function generates a random migratory network
}
