% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomLOOP.R
\name{randomLOOP}
\alias{randomLOOP}
\title{Generate a random network}
\usage{
randomLOOP(
  pop = 1e+05,
  toplot = TRUE,
  nbreeding = 5,
  nwintering = 10,
  npre = 20,
  npost = 20,
  mean_dist = 2000,
  sd_dist = 1000
)
}
\arguments{
\item{pop}{population size flowing through network}

\item{toplot}{TRUE/FALSE to determine whether the output is plotted or not}

\item{nbreeding}{number of breeding sites.}

\item{nwintering}{number of nonbreeding residency/wintering sites.}

\item{npre}{number of sites used duting pre-breeding (north) migration.}

\item{npost}{number of sites used duting post-breeding (south) migration.}

\item{mean_dist}{mean distance an animal can travel}

\item{sd_dist}{standard deviation of the distance an animal can travel}
}
\value{
a list containting the network which was randomly generated,
the tracks that were randomly generated, and the sites that were randomly generated for animals to use.
}
\description{
This function generates a random migratory network
}
\examples{
par(mfrow=c(1,1))
randomLOOP(toplot=TRUE)
randomLOOP(nbreeding = 3,
           nwintering = 3,
           npre = 3,
           npost = 3,
           toplot=TRUE)


}
