% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRIVER.R
\name{randomRIVER}
\alias{randomRIVER}
\title{Generate a random network}
\usage{
randomRIVER(nsites = 50, pop = 1e+05, toplot = TRUE, nbreeding = 8,
  minforks = 2, maxforks = 5)
}
\arguments{
\item{nsites}{number of sites to be generated in the network}

\item{toplot}{TRUE/FALSE to determine whether the output is plotted or not}

\item{nbreeding}{number of breeding sites. If all branch ends should be breeding sites then use "ALL"}

\item{minforks}{minimum number of forks in a river branch}

\item{maxforks}{maximum number of forks in a river branch}
}
\value{
a list containting the network which was randomly generated,
the tracks that were randomly generated, and the sites that were randomly generated for animals to use.
}
\description{
This function generates a random migratory network
}
\examples{
par(mfrow=c(1,1))
randomRIVER(nsites = 20, toplot=T)
randomRIVER(nsites = 50, toplot=T, nbreeding="ALL")


}
