% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortenNET.R
\name{shortenNET}
\alias{shortenNET}
\title{Generate a random network}
\usage{
shortenNET(network, from = "Ssupersink", to = "Nsupersink")
}
\arguments{
\item{network}{the network to be shortened}

\item{from}{the site for the network to be shortened from}

\item{to}{the site for the network to be shortened to}
}
\value{
a list containting the network which was randomly generated,
the tracks that were randomly generated, and the sites that were randomly generated for animals to use.
}
\description{
This function bridges one wintering node to another (rather than going through a supersink)
}
\examples{
par(mfrow=c(1,1))
net <- randomEIGHT(toplot=FALSE)$network
net <- shortenNET(net, from = "Ssupersink", to = "Nsupersink")

}
