% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvCALC.R
\name{curvCALC}
\alias{curvCALC}
\title{Generate a random network}
\usage{
curvCALC(x, y, digits = 3)
}
\arguments{
\item{x}{value for x axis, typically number or proportion of sites lost}

\item{y}{value for y axis, typically population lost or proportion of population lost}

\item{digits}{number of digits to round the result to}
}
\value{
a list containting the network which was randomly generated,
the tracks that were randomly generated, and the sites that were randomly generated for animals to use.
}
\description{
This function generates a random migratory network
}
\examples{

pop=100000
rand_net = randomNET(nsites=100,nbreeding=3, nwintering=3,pop=pop,
                     mean_dist = 8000, sd_dist=9000, toplot=FALSE)

# priotise sites according to flow through network
prioritisation = prioritiseFLOW(rand_net$network, rand_net$sites, toplot=FALSE, returnNET = FALSE)

y=(prioritisation$prioritisation$Pop_Flow/pop)
x=((1:length(prioritisation$prioritisation$Pop_Flow))/
    length(prioritisation$prioritisation$Pop_Flow))


# calculate area under curve
AUC = curvCALC(x,y)
par(mfrow=c(1,1), mar=c(4,4,1,1))
plot(x,y,type="o",pch=20, main = AUC)

}
