% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRIVER.R
\name{randomRIVER}
\alias{randomRIVER}
\title{Generate a random network}
\usage{
randomRIVER(
  nwintering = 4,
  nbreeding = 10,
  nstop = 40,
  pop = 1e+05,
  toplot = TRUE,
  minforks = 2,
  maxforks = 5,
  anadromous = TRUE
)
}
\arguments{
\item{nwintering}{number of nonbreeding residency/wintering sites.}

\item{nbreeding}{number of breeding sites.}

\item{nstop}{number of sites used duting migratory stopover}

\item{pop}{population size flowing through network}

\item{toplot}{TRUE/FALSE to determine whether the output is plotted or not}

\item{minforks}{minimum number of forks in a river branch}

\item{maxforks}{maximum number of forks in a river branch}

\item{anadromous}{TRUE or FALSE. If TRUE, then population breeds at the tips of network branches, otheriwe, it breeding "at sea"}
}
\value{
a list containting the network which was randomly generated,
the tracks that were randomly generated, and the sites that were randomly generated for animals to use.
}
\description{
This function generates a random migratory network
}
\examples{
par(mfrow=c(1,1))
randomRIVER(nbreeding = 6, toplot=TRUE)

}
